Move functions that will be common to resume-only in parent module
